Move matroid constructor
From `constructor.py` to within the `Matroid` class in `matroid.pyx`.
