+++ src/sage/matroids/all.py
@@ -6,7 +6,7 @@ from sage.misc.namespace_package import install_doc
 install_doc(__package__, __doc__)
 
 from sage.misc.lazy_import import lazy_import
-lazy_import('sage.matroids.constructor', 'Matroid')
+lazy_import('sage.matroids.matroid', 'Matroid')
 lazy_import('sage.matroids', 'matroids_catalog', 'matroids')
 del lazy_import
 del install_doc
